extend ring buffer to 1024 bytes

+++ app/isrRingBuffer.c
@@ -2,5 +2,5 @@
 
 uint8_t isrRingBuffer[ISR_RING_BUFFER_LENGTH];
 
-volatile uint8_t isrRingBufferWritePtr = 0;
-volatile uint8_t isrRingBufferReadPtr = 0;
+volatile uint16_t isrRingBufferWritePtr = 0;
+volatile uint16_t isrRingBufferReadPtr = 0;
